refactor: bootstrap uses PluginManager; update manage.ts import
